add piping test

+++ object_detection/detect.py
@@ -20,7 +20,7 @@ class Detect():
             for box in boxes:
                 _box = box.xyxy[0]  # get box coordinates in (top, left, bottom, right) format
                 size = (abs(_box[0].item() - _box[2].item()) + abs(_box[1].item() - _box[3]).item()) / 2
-                position = [(_box[0].item() + _box[2].item()) / 2 , (_box[1].item() + _box[3].item()) / 2]
+                position = [int((_box[0].item() + _box[2].item()) / 2) , int((_box[1].item() + _box[3].item()) / 2)]
                 class_num = int(box.cls.item())
                 if class_num == 0:
                     name = "bent"
